refactor(config): fix 5 layering violations in config module
Fixed upward layering violations in vibe3.config by relocating pure models
and extracting service-dependent helpers:

1. Moved adapter_manifest.py (pure Pydantic models) from config to models
2. Removed TYPE_CHECKING import of ExecutionRole from role_policy.py
3. Extracted block functions from role_policy.py to services/role_policy_helpers.py
4. Extracted orchestra helpers to services/orchestra_helpers.py
5. Removed ConventionResolver dependency from PolicyResolverMixin

Changes:
- New: models/adapter_manifest.py, services/role_policy_helpers.py, services/orchestra_helpers.py
- Deleted: config/adapter_manifest.py
- Updated 27 import sites and 3 agent prompt builders
- Updated test mocks to patch at new import locations

Verification:
- mypy: PASS (no type errors)
- pytest tests/vibe3/config: PASS (72 tests)
- pytest tests/vibe3/execution: PASS (14 noop gate tests)
- pytest tests/vibe3/services: PASS (10 convention resolver tests)
- No remaining config→services/agents/adapters runtime imports
