+++ src/vibe3/agents/review_prompt.py
@@ -25,6 +25,7 @@ from vibe3.models.review import ReviewRequest
 from vibe3.prompts.context_builder import PromptContextBuilder, make_context_builder
 from vibe3.prompts.manifest import PromptManifest, PromptProvider
 from vibe3.resources.runtime_assets import resolve_runtime_asset
+from vibe3.services.convention_resolver import ConventionResolver
 
 ReviewPromptMode = Literal["first", "retry"]
 
@@ -227,14 +228,19 @@ def _build_review_prompt_providers(
     def review_exit_contract() -> str:
         return build_review_task_section(config.review.review_task)
 
+    resolver = ConventionResolver.from_repo()
+
     return {
         "review.policy": lambda: (
             build_policy_section(policy_path)
-            if (policy_path := config.review.get_policy_file())
+            if (
+                policy_path := config.review.policy_file
+                or resolver.get_policy_path("review")
+            )
             else ""
         ),
         "common.rules": lambda: build_tools_guide_section(
-            config.review.get_common_rules()
+            config.review.common_rules or resolver.get_policy_path("common")
         ),
         "review.snapshot_diff": lambda: build_snapshot_diff_section(
             request.structure_diff
