+++ src/vibe3/config/settings.py
@@ -216,56 +216,16 @@ class AgentConfig(BaseModel):
 
 
 class PolicyResolverMixin:
-    """Mixin providing policy resolution methods.
+    """Mixin holding policy_file and common_rules fields.
 
-    Used by ReviewConfig, PlanConfig, RunConfig to avoid DRY violation.
-    Each class must define a _policy_name attribute and policy_file/common_rules fields.
+    Used by ReviewConfig, PlanConfig, RunConfig to define these fields
+    without duplicating field declarations.
 
-    Note: Uses duck typing for policy_file/common_rules access to avoid
-    Pydantic field interference. Subclasses must declare these as actual fields.
+    Note: Subclasses must declare policy_file and common_rules as actual
+    Pydantic fields. This mixin only provides a common base type.
     """
 
     __slots__ = ()
-
-    def get_policy_file(self) -> str | None:
-        """Get policy file path, using profile resolution if not set.
-
-        Returns explicit policy_file if set, otherwise uses ConventionResolver
-        to determine path based on current profile.
-
-        Returns:
-            Path to policy file, or None if not available for current profile.
-        """
-        # Duck typing: subclasses must declare policy_file field
-        policy_file: str | None = getattr(self, "policy_file", None)
-        if policy_file is not None:
-            return policy_file
-
-        from vibe3.services.convention_resolver import ConventionResolver
-
-        resolver = ConventionResolver.from_repo()
-        # Duck typing: subclasses must declare _policy_name
-        policy_name: str = getattr(self, "_policy_name", "unknown")
-        return resolver.get_policy_path(policy_name)
-
-    def get_common_rules(self) -> str | None:
-        """Get common rules path, using profile resolution if not set.
-
-        Returns explicit common_rules if set, otherwise uses ConventionResolver
-        to determine path based on current profile.
-
-        Returns:
-            Path to common rules, or None if not available for current profile.
-        """
-        # Duck typing: subclasses must declare common_rules field
-        common_rules: str | None = getattr(self, "common_rules", None)
-        if common_rules is not None:
-            return common_rules
-
-        from vibe3.services.convention_resolver import ConventionResolver
-
-        resolver = ConventionResolver.from_repo()
-        return resolver.get_policy_path("common")
 
 
 class ReviewConfig(BaseModel, PolicyResolverMixin):
