+++ src/vibe3/execution/codeagent_runner.py
@@ -69,8 +69,8 @@ class CodeagentExecutionService:
         the next tick does not re-dispatch the same issue.
         """
         from vibe3.clients.github_labels import GhIssueLabelPort
-        from vibe3.config.orchestra_config import get_handoff_state_label
         from vibe3.config.orchestra_settings import load_orchestra_config
+        from vibe3.services.orchestra_helpers import get_handoff_state_label
 
         config = load_orchestra_config()
         handoff_label = get_handoff_state_label(config.supervisor_handoff)
