+++ src/vibe3/models/adapter_manifest.py
