+++ src/vibe3/orchestra/queue_operations.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Callable
 
-from vibe3.config.orchestra_config import get_manager_usernames
 from vibe3.domain.qualify_gate import QualifyGateService
 from vibe3.domain.role_resolver import find_role_for_state
 from vibe3.models.orchestra_config import OrchestraConfig
@@ -18,6 +17,7 @@ from vibe3.orchestra.logging import append_orchestra_event
 from vibe3.orchestra.queue_entry import QueueEntry
 from vibe3.orchestra.queue_ordering import sort_ready_issues
 from vibe3.services.label_utils import should_skip_from_queue
+from vibe3.services.orchestra_helpers import get_manager_usernames
 
 if TYPE_CHECKING:
     from vibe3.clients.github_client import GitHubClient
