+++ src/vibe3/roles/governance.py
@@ -11,7 +11,6 @@ from typing import Any
 from loguru import logger
 
 from vibe3.clients.github_client import GitHubClient
-from vibe3.config.orchestra_config import get_manager_usernames
 from vibe3.config.orchestra_settings import load_orchestra_config
 from vibe3.execution import ExecutionRequest
 from vibe3.execution.execution_role_policy import ExecutionRolePolicyService
@@ -40,6 +39,7 @@ from vibe3.roles.governance_utils import (
     find_material_in_catalog,
     get_governed_issue_numbers,
 )
+from vibe3.services.orchestra_helpers import get_manager_usernames
 
 GOVERNANCE_ROLE = RoleDefinition(
     name="governance",
