+++ src/vibe3/roles/governance_utils.py
@@ -6,9 +6,9 @@ from pathlib import Path
 from typing import Any
 
 from vibe3.clients.github_client import GitHubClient
-from vibe3.config.orchestra_config import get_manager_usernames
 from vibe3.models.orchestra_config import OrchestraConfig
 from vibe3.services.label_utils import normalize_assignees, normalize_labels
+from vibe3.services.orchestra_helpers import get_manager_usernames
 from vibe3.services.orchestra_status_service import (
     IssueStatusEntry,
     format_issue_runtime_line,
